Added PayGate Store and Delete Card API

+++ src/main/java/za/co/dladle/service/WalletService.java
@@ -4,8 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Service;
-import za.co.dladle.exception.UserNotFoundException;
 import za.co.dladle.model.PaymentCard;
+import za.co.dladle.paygate.SingleVaultResponse;
+import za.co.dladle.paygate.StatusNameType;
 import za.co.dladle.session.UserSession;
 import za.co.dladle.util.UserUtility;
 
@@ -27,53 +28,102 @@ public class WalletService {
     @Autowired
     private UserUtility userUtility;
 
-    public void addCard(PaymentCard paymentCard) throws UserNotFoundException {
+    @Autowired
+    private PayGateService payGateService;
+
+    public void addCard(PaymentCard paymentCard) throws Exception {
+
+
+        SingleVaultResponse singleVaultResponse = payGateService.singleVault(paymentCard);
+
+        if (singleVaultResponse.getCardVaultResponse() != null) {
+            String vaultId = singleVaultResponse.getCardVaultResponse().getStatus().getVaultId();
+
+            if (vaultId != null) {
+                Map<String, Object> map = new HashMap<>();
+
+                UserSession userSession = applicationContext.getBean("userSession", UserSession.class);
+
+                String cardNumber = paymentCard.getCardNumber().substring(12);
+
+                Long userId = userUtility.findUserIdByEmail(userSession.getUser().getEmailId());
+                map.put("userId", userId);
+                map.put("nameOnCard", paymentCard.getNameOnCard());
+                map.put("cardNumber", "*************" + cardNumber);
+                map.put("cvvNumber", paymentCard.getCvvNumber());
+                map.put("cardType", paymentCard.getCardType());
+                map.put("vaultId", vaultId);
+                map.put("time", LocalDateTime.now());
+
+                String sql = "INSERT INTO payment_card(user_id, card_update_time, name_on_card, card_number, cvv_number, card_type,vaultid) " +
+                        "VALUES (:userId,:time,:nameOnCard,:cardNumber,:cvvNumber,:cardType,:vaultId) ON CONFLICT DO NOTHING ";
+                this.jdbcTemplate.update(sql, map);
+            }
+        } else throw new Exception("Unable to persist card data");
+    }
+
+    public void updateCard(PaymentCard paymentCard) throws Exception {
         Map<String, Object> map = new HashMap<>();
 
         UserSession userSession = applicationContext.getBean("userSession", UserSession.class);
+
+        deleteCard(userSession.getUser().getEmailId());
+
+        SingleVaultResponse singleVaultResponse = payGateService.singleVault(paymentCard);
+
+        String vaultId = singleVaultResponse.getCardVaultResponse().getStatus().getVaultId();
+
+        String cardNumber = paymentCard.getCardNumber().substring(12);
 
         Long userId = userUtility.findUserIdByEmail(userSession.getUser().getEmailId());
         map.put("userId", userId);
         map.put("nameOnCard", paymentCard.getNameOnCard());
-        map.put("cardNumber", paymentCard.getCardNumber());
-        map.put("expiryDate", paymentCard.getExpiryDate());
+        map.put("cardNumber", "*************" + cardNumber);
         map.put("cvvNumber", paymentCard.getCvvNumber());
         map.put("cardType", paymentCard.getCardType());
         map.put("time", LocalDateTime.now());
+        map.put("vaultId", vaultId);
 
-        String sql = "INSERT INTO payment_card(user_id, card_update_time, name_on_card, card_number, expiry_date, cvv_number, card_type) " +
-                "VALUES (:userId,:time,:nameOnCard,:cardNumber,:expiryDate,:cvvNumber,:cardType) ON CONFLICT DO NOTHING ";
+        String sql = "UPDATE payment_card SET name_on_card=:nameOnCard,card_number=:cardNumber,cvv_number=:cvvNumber,card_type=:cardType,card_update_time=:time,vaultid=:vaultId WHERE user_id=:userId";
         this.jdbcTemplate.update(sql, map);
     }
 
-    public void updateCard(PaymentCard paymentCard) throws UserNotFoundException {
+    public void deleteCard() throws Exception {
         Map<String, Object> map = new HashMap<>();
 
         UserSession userSession = applicationContext.getBean("userSession", UserSession.class);
 
         Long userId = userUtility.findUserIdByEmail(userSession.getUser().getEmailId());
-        map.put("userId", userId);
-        map.put("nameOnCard", paymentCard.getNameOnCard());
-        map.put("cardNumber", paymentCard.getCardNumber());
-        map.put("expiryDate", paymentCard.getExpiryDate());
-        map.put("cvvNumber", paymentCard.getCvvNumber());
-        map.put("cardType", paymentCard.getCardType());
-        map.put("time", LocalDateTime.now());
 
-        String sql = "UPDATE payment_card SET name_on_card=:nameOnCard,card_number=:cardNumber,expiry_date=:expiryDate,cvv_number=:cvvNumber,card_type=:cardType,card_update_time=:time WHERE user_id=:userId";
-        this.jdbcTemplate.update(sql, map);
+        map.put("userId", userId);
+
+        String sql1 = "SELECT vaultid FROM payment_card WHERE user_id=:userId";
+
+        String vaultId = this.jdbcTemplate.queryForObject(sql1, map, String.class);
+
+        SingleVaultResponse singleVaultResponse = payGateService.deleteVault(vaultId);
+
+        StatusNameType statusName = singleVaultResponse.getDeleteVaultResponse().getStatus().getStatusName();
+
+        if (statusName.equals(StatusNameType.COMPLETED)) {
+            String sql = "DELETE FROM payment_card WHERE user_id=:userId";
+            this.jdbcTemplate.update(sql, map);
+        } else {
+            throw new Exception("Unable to delete card");
+        }
     }
 
-    public void deleteCard() throws UserNotFoundException {
+    private void deleteCard(String emailId) throws Exception {
         Map<String, Object> map = new HashMap<>();
 
-        UserSession userSession = applicationContext.getBean("userSession", UserSession.class);
-
-        Long userId = userUtility.findUserIdByEmail(userSession.getUser().getEmailId());
+        Long userId = userUtility.findUserIdByEmail(emailId);
         map.put("userId", userId);
 
-        String sql = "DELETE FROM payment_card WHERE user_id=:userId";
-        this.jdbcTemplate.update(sql, map);
+        String sql1 = "SELECT vaultid FROM payment_card WHERE user_id=:userId";
+
+        String vaultId = this.jdbcTemplate.queryForObject(sql1, map, String.class);
+
+        payGateService.deleteVault(vaultId);
     }
 
     public PaymentCard viewCard() throws Exception {
@@ -89,8 +139,6 @@ public class WalletService {
         try {
             return this.jdbcTemplate.queryForObject(sql, map, (rs, rowNum) -> new PaymentCard(rs.getString("name_on_card"),
                     rs.getString("card_number"),
-                    rs.getString("expiry_date"),
-                    rs.getString("cvv_number"),
                     rs.getString("card_type")));
         } catch (Exception e) {
             throw new Exception("No Card available for user");
